refactor: portfolio summary sql query (#33)
* Moved logic to sql query

* Added test

* Fixed precision of big decimal

* Fixed

+++ src/main/kotlin/com/example/stocksheet/analytics/dto/PortfolioSummaryResponseDTO.kt
@@ -6,5 +6,5 @@ data class PortfolioSummaryResponseDTO(
     val totalValue: BigDecimal,
     val totalIncome: BigDecimal,
     val investedCapital: BigDecimal,
-    val todayIncome: Long,
+    val todayIncome: BigDecimal,
 )
